dhc6: fail to rebind when the reply is worse than what we asked for
The DHCPv6 servers don't seem to behave consistently when we attempt to rebind
with more IA_PDs than was previously leased. ISC DHCP would just add an the new
IA_PD, while the OpenWRT's odhcpd would not.

That's a bit unfortunate, as it makes requesting an extra prefix from odhcpd
with ISC dhclient unpractical (shortened for clarity):

  $ dhclient -6 -d -lf $PWD/leases eth0 -N
  XMT: Forming Solicit, 0 ms elapsed.
  XMT:  X-- IA_NA 00:3b:dd:9d
  RCV: Advertise message on eth2 from fe80::5054:ff:feba:444e.
  RCV:  X-- IA_NA 00:3b:dd:9d
  ^C
  $ dhclient -6 -d -lf $PWD/leases eth0 -N -P
  XMT: Forming Rebind, 0 ms elapsed.
  XMT:  X-- IA_NA 00:3b:dd:9d
  RCV: Reply message on eth2 from fe80::5054:ff:feba:444e.
  RCV:  X-- IA_NA 00:3b:dd:9d
  RCV:  X-- IA_PD 00:3b:dd:9d
  RCV:  | !-- Status code of no prefix, IA_PD discarded.

Let's not accept such a lease on rebind an solicit a new one instead.